address code review comment
