+++ src/storage/mod.rs
@@ -108,8 +108,9 @@ use crate::{
         mvcc::{MvccReader, PointGetterBuilder},
         txn::{
             commands::{RawAtomicStore, RawCompareAndSwap, TypedCommand},
-            flow_controller::{EngineFlowController, FlowController},
+            flow_controller::{FlowControlType, FlowController},
             scheduler::Scheduler as TxnScheduler,
+            singleton_flow_controller::EngineFlowController,
             Command,
         },
         types::StorageCallbackType,
@@ -237,7 +238,7 @@ impl<E: Engine, L: LockManager, F: KvFormat> Storage<E, L, F> {
         lock_mgr: L,
         concurrency_manager: ConcurrencyManager,
         dynamic_switches: DynamicConfigs,
-        flow_controller: Arc<dyn FlowController + Send + Sync>,
+        flow_controller: Arc<FlowController>,
         reporter: R,
         resource_tag_factory: ResourceTagFactory,
         quota_limiter: Arc<QuotaLimiter>,
@@ -2811,7 +2812,9 @@ impl<E: Engine, L: LockManager, F: KvFormat> TestStorageBuilder<E, L, F> {
                 pipelined_pessimistic_lock: self.pipelined_pessimistic_lock,
                 in_memory_pessimistic_lock: self.in_memory_pessimistic_lock,
             },
-            Arc::new(EngineFlowController::empty()),
+            Arc::new(FlowController::new(FlowControlType::Singleton(
+                EngineFlowController::empty(),
+            ))),
             DummyReporter,
             self.resource_tag_factory,
             Arc::new(QuotaLimiter::default()),
@@ -2839,7 +2842,9 @@ impl<E: Engine, L: LockManager, F: KvFormat> TestStorageBuilder<E, L, F> {
                 pipelined_pessimistic_lock: self.pipelined_pessimistic_lock,
                 in_memory_pessimistic_lock: self.in_memory_pessimistic_lock,
             },
-            Arc::new(EngineFlowController::empty()),
+            Arc::new(FlowController::new(FlowControlType::Singleton(
+                EngineFlowController::empty(),
+            ))),
             DummyReporter,
             ResourceTagFactory::new_for_test(),
             Arc::new(QuotaLimiter::default()),
