+++ tests/failpoints/cases/test_storage.rs
@@ -33,8 +33,10 @@ use tikv::{
         mvcc::{Error as MvccError, ErrorInner as MvccErrorInner},
         test_util::*,
         txn::{
-            commands, flow_controller::EngineFlowController, Error as TxnError,
-            ErrorInner as TxnErrorInner,
+            commands,
+            flow_controller::{FlowControlType, FlowController},
+            singleton_flow_controller::EngineFlowController,
+            Error as TxnError, ErrorInner as TxnErrorInner,
         },
         Error as StorageError, ErrorInner as StorageErrorInner, *,
     },
@@ -253,11 +255,9 @@ fn test_scale_scheduler_pool() {
     let cfg = new_tikv_config(1);
     let kv_engine = storage.get_engine().kv_engine();
     let (_tx, rx) = std::sync::mpsc::channel();
-    let flow_controller = Arc::new(EngineFlowController::new(
-        &cfg.storage.flow_control,
-        kv_engine.clone(),
-        rx,
-    ));
+    let flow_controller = Arc::new(FlowController::new(FlowControlType::Singleton(
+        EngineFlowController::new(&cfg.storage.flow_control, kv_engine.clone(), rx),
+    )));
 
     let cfg_controller = ConfigController::new(cfg.clone());
     let (scheduler, _receiver) = dummy_scheduler();
